# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachmentView": {
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "media_type": {
            "const": "image/png",
            "title": "Media Type",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "asset_id",
          "media_type",
          "byte_size",
          "width",
          "height",
          "sha256",
          "created_at"
        ],
        "title": "AttachmentView",
        "type": "object"
      },
      "AuthErrorView": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "AuthErrorView",
        "type": "object"
      },
...
      "CatalogPatch": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "expected_revision": {
            "maximum": 2000000000.0,
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/Price"
          },
          "published": {
            "title": "Published",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 300,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "expected_revision",
          "published",
          "enabled",
          "is_default",
          "price",
          "reason"
        ],
        "title": "CatalogPatch",
        "type": "object"
      },
      "CatalogView": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/izo__catalog__schemas__ModelView"
            },
            "title": "Models",
            "type": "array"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderView"
            },
            "title": "Providers",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "permissions",
          "providers",
          "models"
        ],
        "title": "CatalogView",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageView": {
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentView"
            },
            "title": "Attachments",
            "type": "array"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "request_id",
          "role",
          "sequence",
          "content",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "MessageView",
        "type": "object"
      },
...
      "Price": {
        "additionalProperties": false,
        "properties": {
          "currency": {
            "const": "RUB",
            "default": "RUB",
            "title": "Currency",
            "type": "string"
          },
          "image_kopeks_per_image": {
            "anyOf": [
              {
                "maximum": 1000000000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Kopeks Per Image"
          },
          "input_kopeks_per_million": {
            "anyOf": [
              {
                "maximum": 1000000000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Kopeks Per Million"
          },
          "output_kopeks_per_million": {
            "anyOf": [
              {
                "maximum": 1000000000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Kopeks Per Million"
          }
        },
        "title": "Price",
        "type": "object"
      },
...
      "ProviderView": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "ProviderView",
        "type": "object"
      },
...
      "RequestCreate": {
        "additionalProperties": false,
        "properties": {
          "attachment_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 5,
            "title": "Attachment Ids",
            "type": "array"
          },
          "model": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "text": {
            "maxLength": 6000,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "text",
          "model"
        ],
        "title": "RequestCreate",
        "type": "object"
      },
      "RequestView": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "deadline_at": {
            "title": "Deadline At",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "thread_id": {
            "format": "uuid",
            "title": "Thread Id",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "thread_id",
          "model",
          "state",
          "error_code",
          "created_at",
          "updated_at",
          "deadline_at"
        ],
        "title": "RequestView",
        "type": "object"
      },
...
      "ThreadCreate": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ThreadCreate",
        "type": "object"
      },
      "ThreadDetail": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageView"
            },
            "title": "Messages",
            "type": "array"
          },
          "thread": {
            "$ref": "#/components/schemas/ThreadView"
          }
        },
        "required": [
          "thread",
          "messages"
        ],
        "title": "ThreadDetail",
        "type": "object"
      },
      "ThreadList": {
        "properties": {
          "threads": {
            "items": {
              "$ref": "#/components/schemas/ThreadView"
            },
            "title": "Threads",
            "type": "array"
          }
        },
        "required": [
          "threads"
        ],
        "title": "ThreadList",
        "type": "object"
      },
      "ThreadView": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ThreadView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "izo__catalog__schemas__ModelView": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "modality": {
            "enum": [
              "text",
              "image"
            ],
            "title": "Modality",
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/Price"
          },
          "price_source": {
            "enum": [
              "unset",
              "admin_manual"
            ],
            "title": "Price Source",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "publishable": {
            "title": "Publishable",
            "type": "boolean"
          },
          "published": {
            "title": "Published",
            "type": "boolean"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "provider",
          "modality",
          "label",
          "published",
          "enabled",
          "is_default",
          "publishable",
          "price_source",
          "updated_at",
          "price"
        ],
        "title": "ModelView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "IZO ASA Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/catalog": {
      "get": {
        "operationId": "read_api_v1_admin_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe catalog error"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe catalog error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe catalog error"
          }
        },
        "summary": "Read",
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/api/v1/admin/catalog/models/{model_id}": {
      "patch": {
        "operationId": "patch_api_v1_admin_catalog_models__model_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe catalog error"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe catalog error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorView"
                }
              }
            },
            "description": "Safe catalog error"
          }
        },
        "summary": "Patch",
        "tags": [
          "admin-catalog"
        ]
      }
    },
...
    "/api/v1/chat/requests/{request_id}": {
      "get": {
        "operationId": "get_request_api_v1_chat_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Request",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/requests/{request_id}/events": {
      "get": {
        "operationId": "events_api_v1_chat_requests__request_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/requests/{request_id}/stop": {
      "post": {
        "operationId": "stop_request_api_v1_chat_requests__request_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Request",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/threads": {
      "get": {
        "operationId": "threads_api_v1_chat_threads_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Threads",
        "tags": [
          "chat"
        ]
      },
      "post": {
        "operationId": "create_thread_api_v1_chat_threads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Thread",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/threads/{thread_id}": {
      "get": {
        "operationId": "thread_api_v1_chat_threads__thread_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Thread",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/threads/{thread_id}/requests": {
      "post": {
        "operationId": "create_request_api_v1_chat_threads__thread_id__requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Request",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}